+++ packages/pluggableWidgets/datagrid-web/src/components/Row.tsx
@@ -5,6 +5,7 @@ import { SelectActionsService } from "@mendix/widget-plugin-grid/main";
 import { CheckboxCell } from "./CheckboxCell";
 import { DataCell } from "./DataCell";
 import { SelectorCell } from "./SelectorCell";
+import { getRowAriaSelected } from "../features/row-interaction/get-row-aria-selected";
 import { EventsController } from "../typings/CellComponent";
 import { GridColumn } from "../typings/GridColumn";
 
@@ -24,7 +25,7 @@ export interface RowProps {
 export function Row(props: RowProps): ReactElement {
     const { selectActions, totalRows, eventsController } = props;
     const selected = selectActions.isSelected(props.item);
-    const ariaSelected = selectActions.selectionType === "None" ? undefined : selected;
+    const ariaSelected = getRowAriaSelected(selectActions.selectionType, selected, props.checkboxColumnEnabled);
     const borderTop = props.index === 0;
 
     return (